Add LatestExtractDatesForHotelsDataService
The rates table stores extract date in the middle of the row key, which
prevents efficiently determining the latest extract per hotel. Bigtable
doesn’t support ordering or fast reverse scans in this layout, so we
maintain a separate lookup table that stores the latest extract timestamp
per hotel for fast reads.